add typescript + refactor deploy script (wip)
