+++ package.json
@@ -11,6 +11,7 @@
 		"test": "hardhat test --max-memory 4096 --show-stack-traces",
 		"deploy-core-localhost": "npx hardhat deploy-core-localhost --network localhost --show-stack-traces",
 		"deploy-core-goerli": "npx hardhat deploy-core-goerli --network goerli --show-stack-traces",
+		"deploy-core-arbitrum-goerli": "npx hardhat deploy-core-arbitrum-goerli --network arbitrum_goerli --show-stack-traces",
 		"deploy-core-mainnet": "npx hardhat deploy-core-mainnet --network mainnet --show-stack-traces"
 	},
 	"keywords": [],
@@ -54,7 +55,9 @@
 		"prettier-plugin-solidity": "^1.1.3",
 		"readline-sync": "^1.4.10",
 		"solidity-coverage": "^0.8.2",
+		"ts-node": "^10.9.1",
 		"typechain": "^8.1.1",
+		"typescript": "^5.1.3",
 		"web3": "^1.9.0"
 	}
 }
