+++ scripts/deployment/config/localhost.ts
@@ -1,5 +1,5 @@
-const { utils } = require("ethers")
-const toEther = val => utils.parseEther(String(val))
+import { BigNumber, utils } from "ethers"
+const toEther = (val: any): BigNumber => utils.parseEther(String(val))
 
 const OUTPUT_FILE = "./scripts/deployment/output/localhost.json"
 const TX_CONFIRMATIONS = 1
@@ -9,14 +9,12 @@ const CONTRACT_UPGRADES_ADMIN = "0xf99C0eDf98Ed17178B19a6B5a0f3B58753300596"
 const SYSTEM_PARAMS_ADMIN = "0xf99C0eDf98Ed17178B19a6B5a0f3B58753300596"
 const TREASURY_WALLET = "0xf99C0eDf98Ed17178B19a6B5a0f3B58753300596"
 
-// Core Contracts Config ----------------------------------------------------------------------------------------------
-
 const COLLATERAL = [
 	{
 		name: "wETH",
 		address: "0x1A0A7c9008Aa351cf8150a01b21Ff2BB98D70D2D",
 		oracleAddress: "0xE8BAde28E08B469B4EeeC35b9E48B2Ce49FB3FC9",
-		oraclePriceDeviation: toEther(0.25),
+		oracleTimeoutMinutes: 1440,
 		oracleIsEthIndexed: false,
 		MCR: toEther(1.25),
 		CCR: toEther(1.5),
@@ -26,17 +24,10 @@ const COLLATERAL = [
 	},
 ]
 
-// Grvt Contracts Config ----------------------------------------------------------------------------------------------
-
-const GRVT_BENEFICIARIES = {
-	"0x19596e1D6cd97916514B5DBaA4730781eFE49975": 1_000_000,
-}
-
 module.exports = {
 	COLLATERAL,
 	CONTRACT_UPGRADES_ADMIN,
 	ETHERSCAN_BASE_URL,
-	GRVT_BENEFICIARIES,
 	OUTPUT_FILE,
 	SYSTEM_PARAMS_ADMIN,
 	TREASURY_WALLET,
